requirements.txt
Django>=5.0,<6.0
djangorestframework  
psycopg2-binary
gunicorn